Added commandline switch for log path
