+++ DataTongs/Log4Net.config
@@ -23,7 +23,7 @@
             </layout>
         </appender>
         <appender name="ProgressLog" type="log4net.Appender.FileAppender">
-            <file value="DataTongs - Progress.log" />
+            <file type="log4net.Util.PatternString" value="%property{LogPath}/DataTongs - Progress.log" />
             <appendToFile value="false" />
             <layout type="log4net.Layout.PatternLayout">
                 <conversionPattern value="%date - %message%newline" />
@@ -34,7 +34,7 @@
 		    <appender-ref ref="ProgressLog" />
 	    </logger>
 	    <appender name="ErrorLog" type="log4net.Appender.FileAppender">
-		    <file value="DataTongs - Errors.log" />
+		    <file type="log4net.Util.PatternString" value="%property{LogPath}/DataTongs - Errors.log" />
 			<appendToFile value="false" />
 		    <layout type="log4net.Layout.PatternLayout">
 			    <conversionPattern value="%date - %message%newline" />
