+++ SchemaQuench/Program.cs
@@ -10,8 +10,7 @@ public static class Program
 {
     public static void Main(string[] args)
     {
-        if (CommandLineParser.ContainsSwitch("v") || CommandLineParser.ContainsSwitch("ver") || CommandLineParser.ContainsSwitch("version")) ShowVersionAndExit();
-        if (CommandLineParser.ContainsSwitch("?") || CommandLineParser.ContainsSwitch("h") || CommandLineParser.ContainsSwitch("help")) ShowHelpAndExit();
+        CommandLineParser.HandleCommonSwitches("SchemaQuench");
 
         var skipKindlingForge = args.Length > 0 && args[0] == "SkipKindlingForge";
         AppDomain.CurrentDomain.UnhandledException += UnhandledException;
@@ -23,21 +22,5 @@ public static class Program
     public static void UnhandledException(object sender, UnhandledExceptionEventArgs e)
     {
         LogBackup.UnhandledExceptionLogger("SchemaQuench", e);
-    }
-
-    public static void ShowVersionAndExit()
-    {
-        var assembly = Assembly.GetEntryAssembly() ?? Assembly.GetExecutingAssembly();
-        Console.WriteLine($"SchemaQuench - Version: {FileVersionInfo.GetVersionInfo(assembly.Location).FileVersion}");
-        EnvironmentWrapper.GetFromFactory().Exit(0);
-    }
-
-    public static void ShowHelpAndExit()
-    {
-        Console.WriteLine("SchemaQuench.exe [<command>]");
-        Console.WriteLine("  --version                Show the program version");
-        Console.WriteLine("  --ConfigFile:<filepath>  Path and file name of the config file. The default is appsettings.json in the current path.");
-        Console.WriteLine("  --help                   Show the command line options");
-        EnvironmentWrapper.GetFromFactory().Exit(0);
     }
 }
